+++ next-docs/pages/core/search.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Preview from '../../components/codePreview/Preview';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
+import getComponent from '../../components/getComponent';
 import PropsTable from '../../components/PropsTable';
 import Default from '../../public/examples/search/Default';
 import Modal from '../../public/examples/search/Modal';
@@ -8,13 +9,11 @@ import useExamples from '../../utils/useExamples';
 
 const Example = () => {
   const examples = useExamples('search');
+  const { name, text, image } = getComponent('Search');
   return (
     <>
-      <ComponentPageDescription title="Search" isInProgress>
-        <p>
-          Search enables users to specify a word or a phrase to find relevant
-          pieces of content without the use of navigation.
-        </p>
+      <ComponentPageDescription title={name} image={image} isInProgress>
+        <p>{text}</p>
       </ComponentPageDescription>
       <Preview
         title="Default"