+++ packages/select/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @heathmont/moon-select
 
+## 9.26.3
+
+### Patch Changes
+
+- Update docs
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.26.3
+  - @heathmont/moon-themes@9.26.3
+  - @heathmont/moon-utils@9.26.3
+
+## 9.26.2
+
+### Patch Changes
+
+- bump v9.26.2
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.26.2
+  - @heathmont/moon-themes@9.26.2
+  - @heathmont/moon-utils@9.26.2
+
 ## 9.26.1
 
 ### Patch Changes