+++ workspaces/tables/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @heathmont/moon-table-tw
 
+## 9.26.3
+
+### Patch Changes
+
+- Update docs
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.26.3
+
+## 9.26.2
+
+### Patch Changes
+
+- bump v9.26.2
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.26.2
+
 ## 9.26.1
 
 ### Patch Changes